+++ tezos/plugin-js/package.json
@@ -27,17 +27,18 @@
     "@taquito/michel-codec": "12.0.0",
     "@taquito/michelson-encoder": "12.0.0",
     "@taquito/utils": "12.0.0",
-    "@web3api/core-js": "0.0.1-prealpha.61",
+    "@web3api/core-js": "0.0.1-prealpha.76",
     "@temple-wallet/dapp": "6.1.0",
     "axios": "0.25.0"
   },
   "devDependencies": {
     "@types/jest": "26.0.8",
-    "@web3api/client-js": "0.0.1-prealpha.61",
-    "@web3api/ens-plugin-js": "0.0.1-prealpha.61",
-    "@web3api/ipfs-plugin-js": "0.0.1-prealpha.61",
-    "@web3api/test-env-js": "0.0.1-prealpha.61",
-    "@web3api/cli": "0.0.1-prealpha.61",
+    "@blockwatch-cc/tezos-test-env": "0.0.1-prealpha.63",
+    "@web3api/client-js": "0.0.1-prealpha.76",
+    "@web3api/ens-plugin-js": "0.0.1-prealpha.76",
+    "@web3api/ipfs-plugin-js": "0.0.1-prealpha.76",
+    "@web3api/test-env-js": "0.0.1-prealpha.76",
+    "@web3api/cli": "0.0.1-prealpha.76",
     "jest": "26.6.3",
     "js-sha3": "0.8.0",
     "rimraf": "3.0.2",
