+++ tezos/plugin-js/src/common/TezosConfig.ts
@@ -1,0 +1,6 @@
+import { ConnectionConfigs } from "./Connection";
+
+export interface TezosConfig {
+  networks: ConnectionConfigs;
+  defaultNetwork?: string;
+}
