--- tezos/plugin-js/src/utils.ts
@@ -1,3 +1,0 @@
-export function isValidOperationHash(hash: string): boolean {
-    return hash.length === 51 && hash.startsWith('o');
-}
