+++ tezos/plugin-js/web3api.plugin.yaml
@@ -1,3 +1,10 @@
-format: 0.0.1-prealpha.1
+format: 0.0.1-prealpha.2
+name: Tezos
 language: plugin/typescript
-schema: ./schema.graphql
+modules:
+  query:
+    schema: ./src/query/schema.graphql
+    module: ./src/query/index.ts
+  mutation:
+    schema: ./src/mutation/schema.graphql
+    module: ./src/mutation/index.ts
